refactor: 유틸리티 폴더에 있는 모듈을 라이브러리로 호이스팅
