+++ library/decorators/jwt-user-Id.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 
-import { TokenPayload } from '@app/utils/token/types/token.type';
+import { TokenPayload } from '@app/library/token/types/token.type';
 
 export const JwtUserId = createParamDecorator(
   (_: unknown, ctx: ExecutionContext): TokenPayload => {
