+++ library/guards/auth/strategy/jwt-access-token.strategy.ts
@@ -3,7 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 
-import { TokenPayload } from '@app/utils/token/types/token.type';
+import { TokenPayload } from '@app/library/token/types/token.type';
 
 @Injectable()
 export class JwtAccessTokenStrategy extends PassportStrategy(
