--- library/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './constants/constants';
-export * from './date/date';
