+++ src/comments/comments.decorator.ts
@@ -1,7 +1,7 @@
 import { applyDecorators, Delete, Get, Post, Put } from '@nestjs/common';
 import { ApiOkResponse } from '@nestjs/swagger';
 
-import { JwtAccessTokenGuard } from '@app/utils/guards/auth/jwt-access-token.guard';
+import { JwtAccessTokenGuard } from '@app/library/guards/auth/jwt-access-token.guard';
 import { GetCommentsResponseDto } from '@api/comments/dto/get-comments.dto';
 
 export const CreateComments = () =>
