+++ src/posts/like/posts-like.controller.ts
@@ -6,7 +6,7 @@ import {
   UnsetLike,
 } from '@api/posts/like/posts-like.decorator';
 import { PostsLikeService } from '@api/posts/like/posts-like.service';
-import { JwtUserId } from '@app/utils/decorators/jwt-user-Id.decorator';
+import { JwtUserId } from '@app/library/decorators/jwt-user-Id.decorator';
 
 import {
   HasLikedRequestDto,
@@ -14,7 +14,7 @@ import {
 } from '@api/posts/like/dto/has-liked.dto';
 import { SetLikedRequestDto } from '@api/posts/like/dto/set-like.dto';
 import { UnsetLikedRequestDto } from '@api/posts/like/dto/unset-like.dto';
-import { TokenPayload } from '@app/utils/token/types/token.type';
+import { TokenPayload } from '@app/library/token/types/token.type';
 import { unauthorizedUser } from '@api/users/auth/errors/users-auth.error';
 
 @Controller('posts/like')
