+++ src/posts/posts.controller.ts
@@ -12,11 +12,11 @@ import {
   GetPostDetail,
   GetPosts,
 } from '@api/posts/posts.decorator';
-import { JwtUserId } from '@app/utils/decorators/jwt-user-Id.decorator';
+import { JwtUserId } from '@app/library/decorators/jwt-user-Id.decorator';
 
 import { CreatePostRequestBodyDto } from '@api/posts/dto/create-post.dto';
 import { unauthorizedUser } from '@api/users/auth/errors/users-auth.error';
-import { TokenPayload } from '@app/utils/token/types/token.type';
+import { TokenPayload } from '@app/library/token/types/token.type';
 import {
   GetPostDetailRequestQueryDto,
   GetPostDetailResponseDto,
