+++ src/posts/posts.decorator.ts
@@ -1,7 +1,7 @@
 import { applyDecorators, Get, Post } from '@nestjs/common';
 import { ApiOkResponse } from '@nestjs/swagger';
 
-import { JwtAccessTokenGuard } from '@app/utils/guards/auth/jwt-access-token.guard';
+import { JwtAccessTokenGuard } from '@app/library/guards/auth/jwt-access-token.guard';
 
 import { GetPostDetailResponseDto } from '@api/posts/dto/get-post-detail.dto';
 import { GetPostsResponseDto } from '@api/posts/dto/get-posts-detail.dto';
