+++ src/users/auth/users-auth.repository.ts
@@ -3,9 +3,9 @@ import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import { usersAuth } from '@prisma/client';
 
-import { now } from '@app/utils/';
+import { now } from '@app/library/date';
 
-import { TokenPayload } from '@app/utils/token/types/token.type';
+import { TokenPayload } from '@app/library/token/types/token.type';
 import { PrismaConnection } from '@app/library/prisma/type/prisma.type';
 
 @Injectable()
