+++ src/users/users.module.ts
@@ -6,11 +6,11 @@ import { PrismaModule } from '@app/library/prisma';
 import { UsersAuthController } from '@api/users/auth/users-auth.controller';
 import { UsersRepository } from '@api/users/users.repository';
 import { UsersAuthRepository } from '@api/users/auth/users-auth.repository';
-import { GithubStrategy } from '@app/utils/guards/auth/strategy/github.strategy';
+import { GithubStrategy } from '@app/library/guards/auth/strategy/github.strategy';
 import { UsersAuthService } from '@api/users/auth/users-auth.service';
 import { UsersService } from '@api/users/users.service';
-import { CookieService } from '@app/utils/cookie/cookie.service';
-import { JwtAccessTokenStrategy } from '@app/utils/guards/auth/strategy/jwt-access-token.strategy';
+import { CookieService } from '@app/library/cookie/cookie.service';
+import { JwtAccessTokenStrategy } from '@app/library/guards/auth/strategy/jwt-access-token.strategy';
 import { UsersController } from '@api/users/users.controller';
 import { UsersSettingRepository } from '@api/users/setting/users-setting.repository';
 import { UsersSettingService } from '@api/users/setting/users-setting.service';
